fix(react-native): prevent app crashes while multipart uploading
